Added check for maximum key lengths.
Rationale: not all clients support huge key sizes.

RSA is limited to 4096-bit as there is little reason to use a key
size greater than 4096.
DSA is limited to 1024 as most clients don't support anything different. This effectively means
that check #10013 and #10011 are at odds with one another with DSA. This is an acceptable situation
as the use of DSA is fairly rare and RSA is preferred over it. Ultimately, if DSA is being used, you
need to make a choice of whether client compatibility or key size is important, or switch to RSA.
ECDSA is limited to P-256 or P-384. This excludes P-521. It's not immediately clear if all clients
support P-521 though it appears 384 is.