Fix overeager finding of signatures.
The new implementation of VisitAll in #22 walked the signature tree
as deep as possible. The problem with this is that counter signatures
from nested signatures will be included as counter signatures from the
root signature.

This change changes the behavior so that VisitAll only follows nesting
of the same type. A new flag on SignatureKind has been added, Universal,
so that the visitation keeps the eager behavior when it's really wanted.